+++ modules/Accounting/functions.inc.php
@@ -112,6 +112,83 @@ function _makePaymentsTextInput( $value, $name )
 	return TextInput( $value, 'values[' . $id . '][' . $name . ']', '', $extra );
 }
 
+
+/**
+ * Make Payments Comments Input
+ * Add Salaries dropdown to reconcile Payment:
+ * Automatically fills the Comments & Amount inputs.
+ *
+ * @since 5.1
+ *
+ * @uses _makePaymentsTextInput()
+ *
+ * @param  string $value Comments value.
+ * @param  string $name  Column name, 'COMMENTS'.
+ *
+ * @return string Text input if not new or if no Salaries found, else Text input & Salaries dropdown.
+ */
+function _makePaymentsCommentsInput( $value, $name )
+{
+	global $THIS_RET;
+
+	$text_input = _makePaymentsTextInput( $value, $name );
+
+	if ( $THIS_RET['ID'] )
+	{
+		return $text_input;
+	}
+
+	// Add Salaries dropdown to reconcile Payment.
+	$salaries_RET = DBGet( "SELECT ID,TITLE,ASSIGNED_DATE,DUE_DATE,AMOUNT
+		FROM ACCOUNTING_SALARIES
+		WHERE STAFF_ID='" . UserStaffID() . "'
+		AND SYEAR='" . UserSyear() . "'
+		ORDER BY ASSIGNED_DATE DESC
+		LIMIT 100" );
+
+	if ( ! $salaries_RET )
+	{
+		return $text_input;
+	}
+
+	$salaries_options = array();
+
+	foreach ( $salaries_RET as $salary )
+	{
+		$salaries_options[ $salary['AMOUNT'] . '|' . $salary['TITLE'] ] = ProperDate( $salary['ASSIGNED_DATE'], 'short' ) .
+			' — ' . Currency( $salary['AMOUNT'] ) .
+			' — ' . $salary['TITLE'];
+	}
+
+	// JS automatically fills the Comments & Amount inputs.
+	ob_start();
+	?>
+	<script>
+		var accountingPaymentsSalariesReconcile = function( amountComments ) {
+			var separatorIndex = amountComments.indexOf( '|' ),
+				amount = amountComments.substring( 0, separatorIndex ),
+				comments = amountComments.substring( separatorIndex + 1 );
+
+			$('#valuesnewAMOUNT').val( amount );
+			$('#valuesnewCOMMENTS').val( comments );
+		};
+	</script>
+	<?php
+	$js = ob_get_clean();
+
+	// Chosen select so we can search Salaries by date, amount, & title.
+	$select_input = ChosenSelectInput(
+		'',
+		'accounting_salaries',
+		'',
+		$salaries_options,
+		'N/A',
+		'onchange="accountingPaymentsSalariesReconcile(this.options[selectedIndex].value);"'
+	);
+
+	return $text_input . ' ' . $js . $select_input;
+}
+
 function _makeSalariesAmount($value,$column)
 {	global $salaries_total;